Add /gif command handling
- Towards Issue #9

- We handle /gif commands by adding the attachment message to a deferred
queue. msgbot posts a chat.command api call with the /gif command to
the user's private direct message channel with msgbot. When msgbot sees
RightGIF's response, it grabs the image_url for the gif and inserts
that URL as the image_url for the attachment in the deferred message queue.
It deletes the attachment from the deferred message queue and posts it
as normal.

- It will probably be more user-friendly to delete the /gif command
posted to the msgbot private DM channel. We'll look into doing this in
the future.