Initial hooks for /help command
So I have the initial functionality for the help command in place, as well
as generic "Hey you fucked up and that's not a command" sort of handling.
What I need to do is revisit this to ensure it's set up and working the
way we want, and then also to go actually fill out the help documentation.